factorial_attempt_2

+++ 0x08-recursion/main.h
@@ -5,6 +5,6 @@ void _puts_recursion(char *s);
 void _putchar(char n);
 void _print_rev_recursion(char *s);
 int _strlen_recursion(char *s);
-int factorial(int n);
+long int factorial(long int n);
 
 #endif
